random test left

+++ lab3/Malloc and Free/mem.c
@@ -31,7 +31,7 @@
 #endif
 //------------------------------
 
-typedef long Align;
+typedef long long Align;
 typedef unsigned uint;
 typedef unsigned long long ULL;
